pyproject.toml
[tool.poetry]
name = "vid2pdf"
version = "1.1.0"
description = "Simple helper utility to convert a video file to PDF image series"
authors = ["sco1 <sco1.git@gmail.com>"]

readme = "README.md"
homepage = "https://github.com/sco1/"
repository = "https://github.com/sco1/vid2pdf"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Typing :: Typed",
]

[tool.poetry.scripts]
vid2pdf = "vid2pdf:main_cli"

[tool.poetry.dependencies]
python = "^3.8.1"
click = "^8.1"
ffmpy = "^0.3"
Pillow = "^9.0"
tqdm = "^4.62"

[tool.poetry.dev-dependencies]
black = "^23.1"
bump2version = "^1.0"
flake8 = "^6.0"
flake8-annotations = "^3.0"
flake8-bugbear = "^23.1"
flake8-docstrings = "^1.7"
flake8-fixme = "^1.1"
isort = "^5.12"
mypy = "^1.0"
pep8-naming = "^0.13"
pre-commit = "^3.0"

[tool.black]
line-length = 100

[tool.isort]
case_sensitive = true
no_lines_before = "LOCALFOLDER"
order_by_type = false
profile = "black"

[tool.mypy]
disallow_incomplete_defs = true
disallow_untyped_calls = true
disallow_untyped_decorators = true
disallow_untyped_defs = true
ignore_missing_imports = true
no_implicit_optional = true
show_error_codes = true
warn_redundant_casts = true
warn_return_any = true
warn_unused_configs = true
warn_unused_ignores = true

[build-system]
requires = ["poetry-core>=1.2"]
build-backend = "poetry.core.masonry.api"
